+++ spec/classes/bitbucket_backup_spec.rb
@@ -10,9 +10,9 @@ describe 'bitbucket' do
           end
 
           context 'install bitbucket backup client with default params' do
-            it 'should deploy bitbucket backup client 2.0.2 from tar.gz' do
-              should contain_archive("/tmp/bitbucket-backup-distribution-#{BACKUP_VERSION}.tar.gz")
-                .with('source' => "https://marketplace.atlassian.com/download/plugins/com.atlassian.stash.backup.client/version/200000200/bitbucket-backup-distribution-#{BACKUP_VERSION}.tar.gz",
+            it 'should deploy bitbucket backup client 2.0.2 from zip' do
+              should contain_archive("/tmp/bitbucket-backup-distribution-#{BACKUP_VERSION}.zip")
+                .with('source' => "https://maven.atlassian.com/content/groups/public/com/atlassian/bitbucket/server/backup/bitbucket-backup-distribution/#{BACKUP_VERSION}/bitbucket-backup-distribution-#{BACKUP_VERSION}.zip",
                       'extract_path' => '/opt/bitbucket-backup',
                       'creates' => "/opt/bitbucket-backup/bitbucket-backup-client-#{BACKUP_VERSION}/lib",
                       'user' => 'atlbitbucket',
@@ -27,7 +27,7 @@ describe 'bitbucket' do
               should contain_file("/opt/bitbucket-backup/bitbucket-backup-client-#{BACKUP_VERSION}")
                 .with('ensure' => 'directory',
                       'owner'  => 'atlbitbucket',
-                      'group'  => 'atlbitbucket').that_requires("Archive[/tmp/bitbucket-backup-distribution-#{BACKUP_VERSION}.tar.gz]")
+                      'group'  => 'atlbitbucket').that_requires("Archive[/tmp/bitbucket-backup-distribution-#{BACKUP_VERSION}.zip]")
 
               should contain_file('/opt/bitbucket-backup/archives')
                 .with('ensure' => 'directory',
@@ -68,8 +68,8 @@ describe 'bitbucket' do
               { :backupclient_version => '99.43.111' }
             end
             it do
-              should contain_archive('/tmp/bitbucket-backup-distribution-99.43.111.tar.gz')
-                .with('source' => 'https://marketplace.atlassian.com/download/plugins/com.atlassian.stash.backup.client/version/200000200/bitbucket-backup-distribution-99.43.111.tar.gz',
+              should contain_archive('/tmp/bitbucket-backup-distribution-99.43.111.zip')
+                .with('source' => 'https://maven.atlassian.com/content/groups/public/com/atlassian/bitbucket/server/backup/bitbucket-backup-distribution/99.43.111/bitbucket-backup-distribution-99.43.111.zip',
                       'extract_path' => '/opt/bitbucket-backup',
                       'creates' => '/opt/bitbucket-backup/bitbucket-backup-client-99.43.111/lib',
                       'user' => 'atlbitbucket',