+++ CHANGELOG.md
@@ -1,10 +1,14 @@
-##2015-01-18 - Release 2.0.2 (Bitbucker Server)
+##2016-03-05 - Releasing #2.0.3
+###Summary
+- Merged bug fix for bitbucket backup client
+
+##2015-01-18 - Release #2.0.2
 ###Summary
 - Resolved issues #1 and #3
 - Updated test cases
 - Merged pull requests
 
-##2015-11-18 - Release 2.0.1 (Bitbucker Server)
+##2015-11-18 - Release #2.0.1
 ###Summary
 - Add Bitbucket Server 4.0+ changes and support.
 - Remove support for installing upstream java mysql connector